Code refactor, readme, assets/code attribution

+++ assets/shaders/src/shader.vert
@@ -28,7 +28,7 @@ layout(std140, set = 1, binding = 0) readonly buffer ObjectBuffer{
 } objectBuffer;
 
 layout(push_constant) uniform constants {
-	vec4 data;
+	ivec4 data;
 	mat4 render_matrix;
 } pc;
